backend/requirements.txt
boto3
pandas
python-dotenv
openpyxl
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
flask
flask-cors

